requirements.txt
Flask==2.0.24
pyserial==3.5
